Here is an example of what to expect
These are problems that would be emailed beforehand. Again, we want a good bit of notification, so there might be a couple of experienced people in your field around. Additionally, there might be people that have done exactly what you want and things can be resolved well ahead  of time.